Add issue handler to add them to a project.
