+++ pkg/webhooks/github/actions/aggregate_releases.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/Masterminds/semver/v3"
 	"github.com/atedja/go-multilock"
-	v3 "github.com/google/go-github/v57/github"
+	v3 "github.com/google/go-github/v72/github"
 	"github.com/metal-stack/metal-lib/pkg/pointer"
 	"github.com/metal-stack/metal-robot/pkg/clients"
 	"github.com/metal-stack/metal-robot/pkg/config"
