+++ pkg/webhooks/github/actions/docs_preview_comment.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"log/slog"
 
-	v3 "github.com/google/go-github/v57/github"
+	v3 "github.com/google/go-github/v72/github"
 	"github.com/metal-stack/metal-robot/pkg/clients"
 	"github.com/metal-stack/metal-robot/pkg/config"
 	"github.com/mitchellh/mapstructure"
