+++ pkg/webhooks/github/actions/repository_maintainers.go
@@ -6,7 +6,7 @@ import (
 	"log/slog"
 	"strings"
 
-	v3 "github.com/google/go-github/v57/github"
+	v3 "github.com/google/go-github/v72/github"
 	"github.com/metal-stack/metal-robot/pkg/clients"
 	"github.com/metal-stack/metal-robot/pkg/config"
 	"github.com/mitchellh/mapstructure"
